fix: Add stricter checking for `open_stream()` mode
Previously, providing the unsupported "xb" mode for an s3:// URI
would result in reading, even though "x" is a write mode.
The semantics of "x" (exclusive creation, never overwrite)
are not implemented for s3:// streams, so it is better
to throw an error if it is specified.

Similarly, write modes are now formally forbidden
for http(s):// streams, and text modes are forbidden
for both and s3:// and http(s):// streams, since these
are not actually supported.

Additionally, scheme checks are now case-insensitive.