fix: Require Python 3.8+
The existing code is not fully compatible with Python versions lower
than 3.8 due to usage of typing.TypedDict and, potentially, library
incompatibilities. It had not been tested on versions of
Python below 3.8 anyway, so we are not ready to guarantee
its correctness on those versions regardless.

Adding support back in later for Python 3.7 is still possible,
as it is not necessarily a breaking change.
Re-adding support for Python 3.6 is unlikely as Python 3.6 itself
is officially end-of-life.
Python 3.7 will similarly reach end-of-life soon (end of June 2023).