fix: Disallow opening files in text mode through `open_stream`
This was only allowed for files on disk anyway,
but `open_stream` is not intended to be used for anything but
binary data streams, so opening a file in text mode is likely
indicative of a mistake in typing out the file mode.
Raising an error in this case should catch those mistakes.