fix(serialization): Add new data version for files with meta tensors
This prevents files containing meta tensors from being misinterpreted
by older versions of tensorizer that do not support them.
This adds a feature_flags bitfield to the tensorizer file header
that currently records whether a file is encrypted or not,
as the previous way of recording encryption wasn't compatible with
new data versions that may only be required part-way through
serialization.