Major refactor of serialization in preparation for 3.0.
_bulk_write() has been broken up into a series of smaller functions that
each perform some operation on the tensors. _WriteSpec carries the
state of each tensor as it gets processed. Many ops are threaded and
tracked as Futures. Dependent steps do future chaining.

The file schema has changed. Immediately following the metadata block is
the full block of all header structures. We may want to combine the
metadata and the header entries, but they are left separate for now for
easier backwards compatibility.

Previously, hash computations included the hash fields themselves,
presumed to be zeroed. This has been changed so the hash computation
segments omit the hash fields themselves.  This breaks compatibility
with 2.x files for now. This will probably get fixed before 3.0 release