feat(serialization): Store relative offsets for tensor data and headers
Instead of storing the absolute file position of header entries
and tensor data, this change stores their offsets relative to the
beginning of their respective segment (e.g. the header segment, or the
data segment). This makes it easier to create and edit headers without
needing to already know where all prior segments end.

Information about absolute offsets are instead stored on a
segment-by-segment basis in a "layout" section of the metadata,
preceding the metadata and header sections.